+++ src/components/atoms/icons/index.ts
@@ -17,4 +17,4 @@ export { default as Download } from './Download';
 export { default as LinkIcon } from './Link';
 export { default as Plus } from './Plus';
 export { default as Pin } from './Pin';
-export { default as Retweet } from './Retweet';
+export { default as Relay } from './Relay';
